refactor(identity): own SIWE formatter model
