+++ docs/architecture.md
@@ -25,13 +25,12 @@ read from that state. Agent status types are not part of the SDK contract.
 
 ## Identity challenges
 
-`types` defines the EIP-4361 message fields shared by tunnel registration and
-browser/admin wallet login. `portal/identity` owns their formatting and
-verification, while the agent owns the wallet-login lifecycle. Portal emits
-version 1, chain ID 1 messages with LF separators. Nonces use `crypto/rand`;
-the existing secp256k1 and Keccak primitives own EIP-191 personal-sign hashing
-and signer recovery. Personal-sign signatures use `r || s || v`, with `v` in
-`0/1` or `27/28`.
+`portal/identity` owns the EIP-4361 message model, formatting, and verification
+shared by tunnel registration and browser/admin wallet login, while the agent
+owns the wallet-login lifecycle. Portal emits version 1, chain ID 1 messages
+with LF separators. Nonces use `crypto/rand`; the existing secp256k1 and Keccak
+primitives own EIP-191 personal-sign hashing and signer recovery. Personal-sign
+signatures use `r || s || v`, with `v` in `0/1` or `27/28`.
 
 The server retains the exact challenge message, expected address, and expiry.
 Verification requires an exact message match, an unexpired challenge, and a
